+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fixed: Remove empty Last line in TXT and XLSX
 * Changed: {headers:true} now is default option
 * Fixed: option.joinstar flag for SELECT * FROM a,b
+* Added: EXP() function, ^ now is XOR, ~ binary NOT
 
 ### 0.2.2 "Mitilini" (15.01.2016)
 * Fix: SELECT can use functions from GROUP BY list